move breez db to its own package
