+++ db/log.go
@@ -1,0 +1,12 @@
+package db
+
+import (
+	"github.com/breez/lightninglib/daemon"
+	"github.com/btcsuite/btclog"
+)
+
+var log btclog.Logger
+
+func init() {
+	log = daemon.BackendLog().Logger("BRUI")
+}
