+++ init.go
@@ -15,6 +15,7 @@ import (
 
 	breezservice "github.com/breez/breez/breez"
 	"github.com/breez/breez/data"
+	"github.com/breez/breez/db"
 	"github.com/breez/breez/doubleratchet"
 	"github.com/breez/breez/lightningclient"
 	"github.com/breez/lightninglib/daemon"
@@ -73,6 +74,7 @@ var (
 	isReady                      int32
 	started                      int32
 	quitChan                     chan struct{}
+	breezDB                      *db.DB
 )
 
 type Config struct {
@@ -134,14 +136,16 @@ func Start(workingDir string, syncJobMode bool) (chan data.NotificationEvent, er
 		return nil, err
 	}
 
-	if err := openDB(path.Join(appWorkingDir, "breez.db")); err != nil {
+	breezDB, err := db.OpenDB(path.Join(appWorkingDir, "breez.db"))
+	if err != nil {
 		return nil, err
 	}
+
 	if err := doubleratchet.Start(path.Join(appWorkingDir, "sessions_encryption.db")); err != nil {
 		return nil, err
 	}
 	go func() {
-		defer closeDB()
+		defer breezDB.CloseDB()
 		defer doubleratchet.Stop()
 		defer atomic.StoreInt32(&started, 0)
 		defer atomic.StoreInt32(&isReady, 0)
